requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72

langchain-groq==0.3.6
langchain-google-genai==2.1.8

faiss-cpu==1.11.0.post1
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyMuPDF==1.26.3
structlog==25.4.0
docx2txt==0.9
ipykernel==6.30.0
streamlit==1.47.1
pytest==8.4.1
pypdf==5.8.0
python-docx>=0.8.11
python-pptx>=0.6.23
pandas>=2.0.0
openpyxl>=3.1.0
cfn-lint
unstructured[all-docs]>=0.15.0
xlrd==1.2.0
pytest-cov==4.1.0
coverage==7.4.1
deepeval==2.1.5
-e .
